Fixes for getenv, still it doesn't work.
pdksh added for tests.

Note: pdksh Makefile still doesn't work for me (rminnich)

+++ cmd/pdksh/tree.c
@@ -430,7 +430,7 @@ vfptreef(shf, indent, fmt, va)
 
 	while ((c = *fmt++))
 	    if (c == '%') {
-		register int32_t n;
+		register long n;
 		register char *p;
 		int neg;
